Trained Convolution network on Binary dataset
Created a 3 layered sequential model, with a convolution layer, followed by RELU activation function and a max pooling layer which gave a training accuracy of 85.37% and 80.02% on validation accuracy with 10 epochs. Of course, with more epochs, the model would be getting better. The split between training set and validation set was 90-10, that could also result in a relatively low accuracy.